python_training/requirements.txt
# Python Training Pipeline Dependencies
# For WakeOn Drowsiness Detection Model

# Core ML/DL
tensorflow>=2.12.0,<2.16.0
keras>=2.12.0
numpy>=1.23.0,<2.0.0
scipy>=1.10.0

# Computer Vision
opencv-python>=4.7.0
mediapipe>=0.10.0
dlib>=19.24.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.2.0
imbalanced-learn>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# TensorFlow Lite Conversion
tensorflow-model-optimization>=0.7.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
